Address HttpStatus Incompatibility (PAYINP-2233)
When building the forbidden response in the task management controller, when the caller does not have permission to do the operation, use the `ResponseEntity.status()` overload that takes an integer rather than the overload that takes a HttpStatus enum instance.

This addresses an issue whereby when the library is used with Spring 6, which changes the `status` method to take a `HttpStatusCode` interface implementation instead, a `NoSuchMethod` exception is raised. This appears due to the Spring 6 release breaking binary compatibility between Spring 5.

This change should allow the library to continue working in both Spring 5 and Spring 6, without the need for a separate Spring 6 compatible version of the library.